+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFG037VVTTK70QJ5CDNYNTC7"
+        "01KFG3PFMDS83H7GJ6A4M5FQ6T"
     )
     r"""The unique identifier of the data source"""