+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFG037W0VPHKGFN9C44V81A9"
+        "01KFG3PFMM9VEAS69ENYW6GBJF"
     )
     r"""The unique identifier of the data source"""